+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/expr/AbstractImageExpressionEvaluator.java
@@ -353,6 +353,7 @@ public abstract class AbstractImageExpressionEvaluator extends ExpressionEvaluat
             case BINNING -> getBinning();
             case BG_MODEL -> bgRemoval.backgroundModel(arguments);
             case COLUMN_BG_MODEL -> bgRemoval.columnBackgroundModel(arguments);
+            case DESCATTER -> bgRemoval.descatter(arguments);
             case SAVE_RAW -> loader.saveRaw(arguments);
             case BLUR -> convolution.blur(arguments);
             case CLAHE -> clahe.clahe(arguments);